Moved progress.js to utils
